repair (and mv) coordinates tests
